chore: convert entity height and width on float

+++ libs/engine/includes/Components/Animation.hpp
@@ -28,8 +28,8 @@ public:
      * @param offset The number of sprites in the animation.
      */
     Animation(
-        size_t textureWidth, size_t textureHeight, size_t entityWidth,
-        size_t entityHeight, size_t offset, size_t duration
+        size_t textureWidth, size_t textureHeight, float entityWidth,
+        float entityHeight, size_t offset, size_t duration
     )
         : _textureWidth(textureWidth)
         , _textureHeight(textureHeight)
@@ -47,8 +47,8 @@ public:
 
     size_t _textureWidth = 0;
     size_t _textureHeight = 0;
-    size_t _entityWidth = 0;
-    size_t _entityHeight = 0;
+    float _entityWidth = 0;
+    float _entityHeight = 0;
     size_t _offset = 1; ///< The number of sprites in the animation.
     size_t _currentOffset = 0; ///< The current index of the animation.
     size_t _duration = 0;
